Split up predict result per 20 ranks
Due to sheer coincidence, it was getting cut off halfway through the
unicode character in #reddit-soccer and not in the test channel. Because
of this, some clients fell back to parsing it as ISOxxxx, which led to
borked output of the entire message. Instead split it up at the rankings
so we cannot cut a utf8 character off (hopefully).

I choose 20 ranks per message, might have to make it fewer per message
if it would happen to crop up again.